Prefer ipv6 if available.
This introduces a go routine that query for AAAA records too
while discovering auth nameservers. This bring to the table the need
to use mutex everywhere the authServers object is accessed.

This may lead to race conditions. The edits in this commit should
have detected them all, but who knows :)